user login, register and authorization for goal router added

+++ backend/controller/goalController.js
@@ -1,18 +1,20 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 
 //@desc Get goals
 //@route Get /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
 
 //@desc Set goals
-//@route Set /api/goals
+//@route Post /api/goals
 //@access Private
 const setGoal = asyncHandler(async (req, res) => {
     if(!req.body.text) {
@@ -22,6 +24,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -34,11 +37,30 @@ const updateGoals = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id)
 
     if(!goal) {
-        res.status(400)
+        res.status(401)
         throw new Error('Goal not found!')
     }
-    // console.log(req.body);
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
+
+    const user = await User.findById(req.user.id)
+
+    // Check if user exists
+    if(!user) {
+        res.status(401)
+        throw new Error('User does not exist.')
+    }
+
+    // Checking if user is the same user is associated with goal user.
+    if(goal.user.toString() !== req.user.id){
+        console.log(user.id)
+        res.status(401)
+        throw new Error('You are not authorised to access the goal.')
+    }
+
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { 
+                            new: true
+                        })
+
+    // debugger;
 
     res.status(200).json(updatedGoal)
 })
@@ -52,7 +74,21 @@ const deleteGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found!')
     }
-    // console.log(goal)
+
+    const user = await User.find({ email: req.user.email })
+
+    // Check if user exists
+    if(!user) {
+        res.status(401)
+        throw new Error('User does not exist.')
+    }
+
+    // Checking if user is the same user is associated with goal user.
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('You are not authorised to access the goal.')
+    }
+
     await goal.deleteOne()
     
     res.status(200).json({ message: `Delete goals ${req.params.id}` })
